+++ gameEngine/actions.js
@@ -1,101 +1,64 @@
-/* eslint-disable consistent-return */
 /* eslint-disable no-param-reassign */
-/* eslint-disable no-unused-expressions */
-/* eslint-disable no-plusplus */
 const { getRandomUpTo } = require('../gameTerminal/randomFunc');
 
 const { card: cardConst, action } = require('../src/constants');
 
-const { waterCard } = require('./specials');
+const {
+    waterCard,
+    bowArrow,
+    turningPotion,
+    clairvoyance,
+    russianOven,
+    skullLantern,
+} = require('./specials');
+
+const { attackOpponent } = require('./attack');
+
+const { moveCardGraveyard } = require('./graveyard');
 
 const {
     getActivePlayer,
     getInActivePlayer,
-    setHealthToZero,
+    getItemId,
     removePanic,
     lastActionChange,
-    damagePlayer,
+    healPlayer,
     deleteCardsShownProperty,
+    giveCardsTo,
+    setDisabledPropertyToDefault,
 } = require('./helpers');
 
-const { moveToGraveyard } = require('./graveyard');
+// function that change turn in the game
+function changeTurn(game) {
+    // active player becomes inactive
+    const pActive = getActivePlayer(game);
+    const pInactive = getInActivePlayer(game);
+    // player's counter set to 0
+    pActive.moveCounter = 0;
+    // inactive player becomes active
+    game.active = pInactive.id;
+    // and run function water if any
+    waterCard(game.players);
+}
 
-function attackShield(player, itemKey, points, opponent) {
-    if (player.item[itemKey].healthCurrent > points) {
-        player.item[itemKey].healthCurrent -= points;
-    } else if (player.item[itemKey].healthCurrent === points) {
-        moveToGraveyard(player, itemKey, opponent);
+function turningPotionEffect(pInactive, pActive, cardId) {
+    if (pActive.turningHand === true) {
+        moveCardGraveyard(pInactive, cardId);
     } else {
-        damagePlayer(player, points - player.item[itemKey].healthCurrent);
-        moveToGraveyard(player, itemKey, opponent);
+        moveCardGraveyard(pActive, cardId);
     }
 }
 
-// TODO: rename to magicMirror/plateMail or not
-// function to reflect half of the damage (or round down to integer) for magicMirror card
-// if active card has 1 attack points then each player gets 1 pnt damage
-// or in case of plateMail func reflect 1 point back to player
-// if active card has 1 attack points then no damage to both players
-function reflect(opponent, player, points) {
-    const itemCard = Object.values(opponent.item)[0].id;
-    let damage;
-    if (itemCard === cardConst.MAGICMIRRORCARD) {
-        points === 1 ? (damage = 1) : (damage = Math.floor(points / 2));
-        opponent.health.current -= damage;
-    }
-    if (itemCard === cardConst.PLATEMAILCARD) {
-        points === 1 ? (damage = 0) : (damage = 1);
-        damage === 0
-            ? opponent.health.current
-            : (opponent.health.current -= points - damage);
-    }
-    if (
-        Object.keys(player.item).length === 0
-    || Object.values(player.item)[0].category !== cardConst.SHIELDCARD
-    ) {
-        player.health.current -= damage;
-    }
-    if (
-        Object.keys(player.item).length === 1
-    && Object.values(player.item)[0].category === cardConst.SHIELDCARD
-    ) {
-        attackShield(player, Object.keys(player.item)[0], damage);
-    }
-    setHealthToZero(player.health.current);
-    setHealthToZero(opponent.health.current);
-}
+// EXPORTS
 
-// function to handle attack points deduction from opponent player health
-// depending on item card if any present
-function attackOpponent(player, opponent, points) {
-    // console.log('attackOpponent ', player, points);
-    let itemCategory;
-    const itemKey = Object.keys(player.item)[0];
-    itemKey ? (itemCategory = player.item[itemKey].category) : null;
-
-    // we check if item holder is not empty and item card does not have shield category
-    const itemLength = Object.keys(player.item).length;
-    if (itemLength === 0 || itemCategory !== cardConst.SHIELDCATEGORY) {
-    // and if item card does not have 'reflect' category (e.g. magicMirror card)
-    // we descrease hero health for attack card points
-        if (itemCategory !== cardConst.REFLECTCATEGORY) {
-            player.health.current -= points;
-        }
-        // we check for special mirror card at opponent item holder with category 'reflect'
-        // and run reflect function if it is present there
-        if (itemCategory === cardConst.REFLECTCATEGORY) {
-            opponent.turningHand === true
-                ? reflect(opponent, player, points)
-                : reflect(player, opponent, points);
-        }
-        if (player.health.current <= 0) {
-            player.health.current = 0;
-        }
-    } else if (itemLength === 1 && itemCategory === cardConst.SHIELDCATEGORY) {
-    // console.log('Were in attack shield');
-    // console.log(itemKey);
-        attackShield(player, itemKey, points, opponent);
+function changeMoveCounter(pActive, card, pInactive) {
+    if (pActive.hand[card] === undefined && pActive.turningHand === false) {
+        pActive.moveCounter += 1;
+        const noClairvoyance = (player) => setTimeout(() => deleteCardsShownProperty(player), 1000);
+        noClairvoyance(pInactive);
+        noClairvoyance(pActive);
     }
+    return pActive;
 }
 
 // we move item card from active player's hand to his item holder
@@ -111,9 +74,11 @@ function moveItem(game, player, key, opponent) {
             if (player.item[key].panic === false) {
                 // console.log('We delete panic!');
                 delete player.item[key].panic;
-                player.turningHand === true
-                    ? removePanic(opponent)
-                    : removePanic(player);
+                if (player.turningHand === true) {
+                    removePanic(opponent);
+                } else {
+                    removePanic(player);
+                }
             }
             // and also we delete the card with active key from player's hand
             delete player.hand[key];
@@ -134,29 +99,6 @@ function moveItem(game, player, key, opponent) {
     lastActionChange(game, action.ITEM);
 }
 
-function changeMoveCounter(pActive, card, pInactive) {
-    if (pActive.hand[card] === undefined && pActive.turningHand === false) {
-        pActive.moveCounter += 1;
-        const noClairvoyance = (player) => setTimeout(() => deleteCardsShownProperty(player), 1000);
-        noClairvoyance(pInactive);
-        noClairvoyance(pActive);
-    }
-    return pActive;
-}
-
-// function that change turn in the game
-function changeTurn(game) {
-    // active player becomes inactive
-    const pActive = getActivePlayer(game);
-    const pInactive = getInActivePlayer(game);
-    // player's counter set to 0
-    pActive.moveCounter = 0;
-    // inactive player becomes active
-    game.active = pInactive.id;
-    // and run function water if any
-    waterCard(game.players);
-}
-
 // function forestMushroom accepts opponent
 // and if opponent get 60% chance, then his/ her cards in hand
 // get panic: true property except one random card
@@ -166,12 +108,13 @@ function forestMushroom(game, opponent) {
     const chance = getRandomUpTo(10, 'chanceforestMushroom');
     const opponentCards = Object.values(opponent.hand);
     if (chance <= 6) {
+        // eslint-disable-next-line no-plusplus
         for (let i = 0; i < Object.keys(opponent.hand).length; i++) {
             opponentCards[i].panic = true;
         }
-        Object.keys(opponent.item).length !== 0
-            ? (Object.values(opponent.item)[0].panic = true)
-            : null;
+        if (Object.keys(opponent.item).length !== 0) {
+            Object.values(opponent.item)[0].panic = true;
+        }
         const index = getRandomUpTo(opponentCards.length, 'indexMushroom');
         opponentCards[index].panic = false;
         if (opponentCards[index].disabled === true) {
@@ -184,11 +127,122 @@ function forestMushroom(game, opponent) {
     return opponentCards;
 }
 
+// checks if opponent item is not empty and if opponent has magicTree card in item
+// if so - function change turn runs after moveCounter === 1, active layer becomes inactive etc.
+function magicTree(game) {
+    const pActive = getActivePlayer(game);
+    const pInactive = getInActivePlayer(game);
+    const itemId = getItemId(pInactive.item);
+    if (pActive.moveCounter === 1 && itemId === cardConst.MAGICTREECARD) {
+        giveCardsTo(pActive);
+        changeTurn(game);
+    }
+}
+
+function playerMoveEnd(pActive, pInactive, game) {
+    // we call function to return disabled cards property to false if any have true
+    setDisabledPropertyToDefault(pActive);
+    // we call function to give cards to players up to 5
+    if (pActive.health.current > 0) {
+        giveCardsTo(pActive);
+    }
+
+    changeTurn(game);
+
+    //  after change of turn,  we check
+    // whether inactive player has in item holder card forest Mushroom with category panic,
+    // then we call function forestMushroom
+    if (Object.keys(pActive.item).length !== 0) {
+        if (Object.values(pActive.item)[0].category === cardConst.PANICCATEGORY) {
+            forestMushroom(game, pInactive, 'afterTurn');
+        }
+    }
+    // we run bowArrow function to check if opponent has bow & arrow card in item
+    // and to supress attack points if any
+    bowArrow(pActive, pInactive);
+}
+
+// eslint-disable-next-line consistent-return
+function pActiveIsTarget(game, activeCard, cardId) {
+    const pActive = getActivePlayer(game);
+    const pInactive = getInActivePlayer(game);
+    switch (activeCard.category) {
+    case cardConst.HEALCATEGORY:
+        healPlayer(pActive, activeCard.points);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.HEAL);
+        if (pActive.turningHand === true) {
+            moveCardGraveyard(pInactive, cardId);
+        } else {
+            moveCardGraveyard(pActive, cardId);
+        }
+        break;
+    case cardConst.ATTACKCATEGORY:
+        break;
+    // if any mistake occurs during game process, player gets error message by default
+    default:
+        return new Error('You are under spell. Wait for redemption!');
+    }
+}
+
+// eslint-disable-next-line consistent-return
+function pInactiveIsTarget(game, activeCard, cardId) {
+    const pActive = getActivePlayer(game);
+    const pInactive = getInActivePlayer(game);
+    switch (activeCard.category) {
+    case cardConst.HEALCATEGORY:
+        break;
+    case cardConst.ATTACKCATEGORY:
+        attackOpponent(pInactive, pActive, activeCard.points);
+        turningPotionEffect(pInactive, pActive, cardId);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.ATACKOPPONENT);
+        break;
+    // if player attacks with card category holdCard we call disableCards function
+    // then move this attack card to gravyeard
+    case cardConst.HOLDCARDCATEGORY:
+        russianOven(pInactive);
+        turningPotionEffect(pInactive, pActive, cardId);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.CHAINS);
+        break;
+    // if player attacks with card category == attackItems, we call attack items function
+    // then move this attack card to gravyeard
+    case cardConst.ATTACKITEMSCATEGORY:
+        skullLantern(game.players);
+        turningPotionEffect(pInactive, pActive, cardId);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.ATTACKITEMS);
+        break;
+    // if player attackes with clairvoyance card, we call showCards function
+    // then move this attack card to gravyeard
+    case cardConst.SHOWCARDSCATEGORY:
+        clairvoyance(pInactive);
+        turningPotionEffect(pInactive, pActive, cardId);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.CLAIRVOYANCE);
+        break;
+    // if player attackes with turningPotion card, we call turningHand function
+    // then move this attack card to gravyeard
+    case cardConst.TURNINGCATEGORY:
+        turningPotion(pActive, pInactive);
+        moveCardGraveyard(pActive, cardId);
+        // after player's act we change lastAction property of the game
+        lastActionChange(game, action.TURNINGPOTION);
+        break;
+
+    // if any mistake occurs during game process, player gets error message by default
+    default:
+        return new Error('You are under spell. Wait for redemption!');
+    }
+}
+
 module.exports = {
-    attackShield,
-    moveItem,
     changeMoveCounter,
-    changeTurn,
+    moveItem,
     forestMushroom,
-    attackOpponent,
+    magicTree,
+    playerMoveEnd,
+    pActiveIsTarget,
+    pInactiveIsTarget,
 };
